Move implementation detail to separate files
