+++ core/include/scipp/core/element/bin_detail.h
