+++ core/test/CMakeLists.txt
@@ -8,7 +8,7 @@ add_executable(
   element_array_test.cpp
   element_array_view_test.cpp
   element_arithmetic_test.cpp
-  element_bin_util_test.cpp
+  element_bin_detail_test.cpp
   element_comparison_test.cpp
   element_event_operations_test.cpp
   element_geometric_operations_test.cpp
